% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting_functions.R
\name{kw.mob}
\alias{kw.mob}
\title{Calculate KW-MOB pseudo-weights}
\usage{
kw.mob(psa_dat, wt, rsp_name, formula, tune_maxdepth, covars, h = NULL,
  krn = "triang", Large = F, rm.s = F)
}
\arguments{
\item{psa_dat}{Dataframe of the combined non-probability and probability sample}

\item{wt}{Name of the weight variable in \code{psa_dat}
(common weights of 1 for non-probability sample, and survey weights for probability sample)}

\item{rsp_name}{Name of the non-probability sample membership indicator in \code{psa_dat}
(1 for non-probability sample units, and 0 for probability sample units)}

\item{formula}{Formula of the propensity model
(see \code{glmtree()} in \code{partykit} package)}

\item{tune_maxdepth}{A vector of values for the tuning parameter maxdepth
(see \code{glmtree()} in \code{partykit} package)}

\item{covars}{A vector of covariate names for standardized mean differences
(SMD; covariate balance) calculation}

\item{h}{Bandwidth parameter
(will be calculated corresponding to kernel function if not specified)}

\item{krn}{Kernel function.
"\code{triang}": triangular density on (-3, 3),
"\code{dnorm}": standard normal density,
"\code{dnorm_t}": truncated standard normal density on (-3, 3).}

\item{Large}{The cohort size is so large that it has to be divided into pieces. Default is \code{FALSE}.}

\item{rm.s}{Remove unmatched survey units or not. Default is \code{FALSE}.}
}
\value{
A list \cr
\code{pswt}: A dataframe including KW pseudo-weights for each tuning parameter setting \cr
\code{smds}: A vector of SMD for each set of KW pseudo-weights \cr
\code{best}: Identifier for the KW pseudo-weights in pswt with the smallest SMD \cr
\code{p_score_c}: A dataframe including propensity scores for non-probability sample units
for each tuning parameter setting \cr
\code{p_score_s}: A dataframe including propensity scores for probability sample units
for each tuning parameter setting
}
\description{
This function computes KW pseudo-weights using model-based recursive partitioning
(\code{glmtree()} in \code{partykit} package) to predict propensity scores.
}
\examples{
# KW-MOB with example data
kwmob <- kw.mob(simu_dat, "wt", "trt",
                "trt_f ~ x1+x2+x3+x4+x5+x6+x7 | x1+x2+x3+x4+x5+x6+x7",
                tune_maxdepth = c(2, 3),
                covars = c("x1","x2","x3","x4","x5","x6","x7"))
# Select KW-MOB pseudo-weights with best covariate balance
kwmob_w <- kwmob$pswt[, kwmob$best]
# Compute weighted mean of y in non-prob data
sum((simu_dat$y[simu_dat$trt == 1]*kwmob_w)/sum(kwmob_w))
}
