% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting_functions.R
\name{kw.wt}
\alias{kw.wt}
\title{Calculate KW pseudo-weights given propensity scores}
\usage{
kw.wt(p_score.c, p_score.s, svy.wt, h = NULL, mtch_v = NULL,
  krn = "triang", Large = F, rm.s = F)
}
\arguments{
\item{p_score.c}{Predicted propensity scores for the non-probability sample}

\item{p_score.s}{Predicted propensity scores for the probability sample}

\item{svy.wt}{A vector of survey weights for the probability sample units}

\item{h}{Bandwidth parameter
(will be calculated corresponding to kernel function if not specified)}

\item{krn}{Kernel function.
"triang": triangular density on (-3, 3),
"dnorm": standard normal density,
"dnorm_t": truncated standard normal density on (-3, 3).}

\item{Large}{The cohort size is so large that it has to be divided into pieces. Default is FALSE.}

\item{rm.s}{Remove unmatched survey units or not. Default is FALSE.}
}
\value{
A list \cr
pswt: KW pseudo-weights \cr
delt.svy: Number of unmatched survey sample units \cr
h: Bandwidth
}
\description{
This function computes KW pseudo-weights based on propensity scores
that are provided by the user.
}
\details{
If there are unmatched survey sample units, the program gives
"The input bandwidth h is too small. Please choose a larger one!"
If rm.s=TRUE, the program deletes unmatched survey sample units, and gives
a warning "records in the prob sample were not used because of a small bandwidth"
If rm.s=FALSE, the program evenly distributes weights of unmatched survey sample units
to all non-probability sample units.
}
