% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwml.R
\docType{package}
\name{KWML}
\alias{KWML}
\alias{KWML-package}
\title{KWML: Boosted Kernel Weighting for R}
\description{
R package for "Boosted Kernel Weighting - Using Statistical Learning to Improve Inference from Nonprobability Samples"
}
