% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simu_dat}
\alias{simu_dat}
\title{Example data set}
\format{A data frame with 4000 rows and 15 variables:
\describe{
  \item{w1, ..., w10}{Covariates}
  \item{y}{Outcome}
  \item{wt}{True weights}
  \item{elig_wt}{Observed weights (= 1 for non-probability sample units)}
  \item{trt_n}{Sample membership indicator (numeric)}
  \item{trt}{Sample membership indicator (factor)}
}}
\usage{
simu_dat
}
\description{
A dataset containing a simulated non-probability (n_np = 2000)
and probability (n_p = 2000) sample.
}
\keyword{datasets}
