% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting_functions.R
\name{ipsw.lg}
\alias{ipsw.lg}
\title{Calculate pseudo-weights using IPSW}
\usage{
ipsw.lg(psa_dat, wt, rsp_name, formula)
}
\arguments{
\item{psa_dat}{Dataframe of the combined non-probability and probability sample}

\item{wt}{Name of the weight variable in psa_dat
(common weights of 1 for non-probability sample, and survey weights for probability sample)}

\item{rsp_name}{Name of the non-probability sample membership indicator in psa_dat
(1 for non-probability sample units, and 0 for probability sample units)}

\item{formula}{Formula of the regression model}
}
\value{
A vector of IPSW pseudo-weights
}
\description{
This function computes IPSW pseudo-weights using logistic regression to
predict propensity scores.
}
