% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting_functions.R
\name{kw.lg}
\alias{kw.lg}
\title{Calculate KW-LG pseudo-weights}
\usage{
kw.lg(psa_dat, wt, rsp_name, formula, h = NULL, krn = "triang",
  Large = F, rm.s = F)
}
\arguments{
\item{psa_dat}{Dataframe of the combined non-probability and probability sample}

\item{wt}{Name of the weight variable in psa_dat
(common weights of 1 for non-probability sample, and survey weights for probability sample)}

\item{rsp_name}{Name of the non-probability sample membership indicator in psa_dat
(1 for non-probability sample units, and 0 for probability sample units)}

\item{formula}{Formula of the regression model}

\item{h}{Bandwidth parameter
(will be calculated corresponding to kernel function if not specified)}

\item{krn}{Kernel function.
"triang": triangular density on (-3, 3),
"dnorm": standard normal density,
"dnorm_t": truncated standard normal density on (-3, 3).}

\item{Large}{The cohort size is so large that it has to be divided into pieces. Default is FALSE.}

\item{rm.s}{Remove unmatched survey units or not. Default is FALSE.}
}
\value{
A list \cr
pswt: KW pseudo-weights \cr
delt.svy: Number of unmatched survey sample units \cr
h: Bandwidth
}
\description{
This function computes KW pseudo-weights using logistic regression to
predict propensity scores.
}
