% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting_functions.R
\name{kw.wt}
\alias{kw.wt}
\title{Calculate KW pseudo-weights given propensity scores}
\usage{
kw.wt(p_score.c, p_score.s, svy.wt, h = NULL, krn = "triang",
  Large = F, rm.s = F)
}
\arguments{
\item{p_score.c}{Predicted propensity scores for the non-probability sample}

\item{p_score.s}{Predicted propensity scores for the probability sample}

\item{svy.wt}{A vector of survey weights for the probability sample units}

\item{h}{Bandwidth parameter
(will be calculated corresponding to kernel function if not specified)}

\item{krn}{Kernel function.
"\code{triang}": triangular density on (-3, 3),
"\code{dnorm}": standard normal density,
"\code{dnorm_t}": truncated standard normal density on (-3, 3).}

\item{Large}{The cohort size is so large that it has to be divided into pieces. Default is \code{FALSE}.}

\item{rm.s}{Remove unmatched survey units or not. Default is \code{FALSE}.}
}
\value{
A list \cr
\code{pswt}: KW pseudo-weights \cr
\code{delt.svy}: Number of unmatched survey sample units \cr
\code{h}: Bandwidth
}
\description{
This function computes KW pseudo-weights based on propensity scores
that are provided by the user.
}
\details{
If there are unmatched survey sample units, the program gives
"The input bandwidth h is too small. Please choose a larger one!"
If \code{rm.s=TRUE}, the program deletes unmatched survey sample units, and gives
a warning "records in the prob sample were not used because of a small bandwidth"
If \code{rm.s=FALSE}, the program evenly distributes weights of unmatched survey sample units
to all non-probability sample units.
}
